image-creation/requirements.txt
Pillow
Flask
opencv-python
numpy